Move splits_from_settings -> settings_gui
